+++ packages/graphql-language-service-interface/src/__tests__/getDiagnostics-test.ts
@@ -12,7 +12,11 @@ import fs from 'fs';
 import { buildSchema, parse, GraphQLSchema } from 'graphql';
 import path from 'path';
 
-import { getDiagnostics, validateQuery, SEVERITY } from '../getDiagnostics';
+import {
+  getDiagnostics,
+  validateQuery,
+  DIAGNOSTIC_SEVERITY,
+} from '../getDiagnostics';
 
 describe('getDiagnostics', () => {
   let schema: GraphQLSchema;
@@ -30,7 +34,7 @@ describe('getDiagnostics', () => {
     expect(error.message).toEqual(
       'Cannot query field "title" on type "Query".',
     );
-    expect(error.severity).toEqual(SEVERITY.ERROR);
+    expect(error.severity).toEqual(DIAGNOSTIC_SEVERITY.Error);
     expect(error.source).toEqual('GraphQL: Validation');
   });
 
@@ -40,9 +44,10 @@ describe('getDiagnostics', () => {
       schema,
     )[0];
     expect(error.message).toEqual(
-      'The field Query.deprecatedField is deprecated. Use test instead.',
+      // eslint-disable-next-line no-useless-escape
+      'The field "Query.deprecatedField" is deprecated. Use test instead.',
     );
-    expect(error.severity).toEqual(SEVERITY.WARNING);
+    expect(error.severity).toEqual(DIAGNOSTIC_SEVERITY.Warning);
     expect(error.source).toEqual('GraphQL: Deprecation');
   });
 
@@ -71,8 +76,11 @@ describe('getDiagnostics', () => {
     );
     expect(errors.length).toEqual(1);
     const error = errors[0];
-    expect(error.message).toEqual('Syntax Error: Expected :, found Name "id"');
-    expect(error.severity).toEqual(SEVERITY.ERROR);
+    expect(error.message).toEqual(
+      // eslint-disable-next-line no-useless-escape
+      'Syntax Error: Expected ":", found Name "id".',
+    );
+    expect(error.severity).toEqual(DIAGNOSTIC_SEVERITY.Error);
     expect(error.source).toEqual('GraphQL: Syntax');
   });