+++ packages/graphql-language-service-interface/src/index.ts
@@ -7,24 +7,15 @@
  *
  */
 
-export {
-  getDefinitionState,
-  getFieldDef,
-  forEachState,
-  objectValues,
-  hintList,
-} from './autocompleteUtils';
+export * from './autocompleteUtils';
 
-export { getAutocompleteSuggestions } from './getAutocompleteSuggestions';
+export * from './getAutocompleteSuggestions';
 
-export {
-  LANGUAGE,
-  getDefinitionQueryResultForFragmentSpread,
-  getDefinitionQueryResultForDefinitionNode,
-} from './getDefinition';
+export * from './getDefinition';
 
-export { getDiagnostics, validateQuery } from './getDiagnostics';
+export * from './getDiagnostics';
+
 export { getOutline } from './getOutline';
 export { getHoverInformation } from './getHoverInformation';
 
-export { GraphQLLanguageService } from './GraphQLLanguageService';
+export * from './GraphQLLanguageService';