+++ packages/graphql-language-service-parser/src/RuleHelpers.ts
@@ -5,12 +5,11 @@
  *  This source code is licensed under the license found in the
  *  LICENSE file in the root directory of this source tree.
  *
- *  @flow
  */
 
 // These functions help build matching rules for ParseRules.
 
-import { Rule, Token } from 'graphql-language-service-types';
+import { Rule, Token } from './types';
 
 // An optional rule.
 export function opt(ofRule: Rule | string): Rule {