+++ packages/graphql-language-service-server/src/index.ts
@@ -7,12 +7,13 @@
  *
  */
 
-export { GraphQLCache, getGraphQLCache } from './GraphQLCache';
-
 export { MessageProcessor } from './MessageProcessor';
 
 export { default as startServer } from './startServer';
 
-export { findGraphQLTags } from './findGraphQLTags';
-
 export { Logger } from './Logger';
+
+export * from './GraphQLCache';
+export * from './parseDocument';
+export * from './findGraphQLTags';
+export * from './stringToHash';