+++ packages/graphql-language-service/src/cli.ts
@@ -27,6 +27,8 @@ const { argv } = yargs
   )
   .help('h')
   .alias('h', 'help')
+  .strict()
+  .recommendCommands()
   .demandCommand(
     1,
     'At least one command is required.\n' +
@@ -81,6 +83,7 @@ const { argv } = yargs
       'Can be one of: stream, node, socket.\n' +
       'Will default to use a node IPC channel for communication.\n',
     type: 'string',
+    default: 'node',
   })
   .option('p', {
     alias: 'port',