+++ packages/graphql-language-service-utils/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.4.0-alpha.6](https://github.com/graphql/graphiql/compare/graphql-language-service-utils@2.4.0-alpha.5...graphql-language-service-utils@2.4.0-alpha.6) (2020-04-10)
+
+**Note:** Version bump only for package graphql-language-service-utils
+
+# [2.4.0-alpha.5](https://github.com/graphql/graphiql/compare/graphql-language-service-utils@2.4.0-alpha.4...graphql-language-service-utils@2.4.0-alpha.5) (2020-04-06)
+
+### Features
+
+- upgrade to graphql@15.0.0 for [#1191](https://github.com/graphql/graphiql/issues/1191) ([#1204](https://github.com/graphql/graphiql/issues/1204)) ([f13c8e9](https://github.com/graphql/graphiql/commit/f13c8e9d0e66df4b051b332c7d02f4bb83e07ffd))
+
+# [2.4.0-alpha.4](https://github.com/graphql/graphiql/compare/graphql-language-service-utils@2.4.0-alpha.3...graphql-language-service-utils@2.4.0-alpha.4) (2020-04-03)
+
+**Note:** Version bump only for package graphql-language-service-utils
+
+# [2.4.0-alpha.3](https://github.com/graphql/graphiql/compare/graphql-language-service-utils@2.4.0-alpha.2...graphql-language-service-utils@2.4.0-alpha.3) (2020-03-20)
+
+**Note:** Version bump only for package graphql-language-service-utils
+
+# [2.4.0-alpha.2](https://github.com/graphql/graphiql/compare/graphql-language-service-utils@2.4.0-alpha.0...graphql-language-service-utils@2.4.0-alpha.2) (2020-03-20)
+
+### Bug Fixes
+
+- exclusions rule for custom rules ([2bdc132](https://github.com/graphql/graphiql/commit/2bdc132abb3b1a0c5ad644e53a2c6e070a2185ce))
+
+### Features
+
+- use new GraphQL Config ([#1342](https://github.com/graphql/graphiql/issues/1342)) ([e45838f](https://github.com/graphql/graphiql/commit/e45838f5ba579e05b20f1a147ce431478ffad9aa))
+
 # [2.4.0-alpha.1](https://github.com/graphql/graphiql/compare/graphql-language-service-utils@2.3.3...graphql-language-service-utils@2.4.0-alpha.1) (2020-01-18)
 
 ### Bug Fixes